JGC-471 - Add output format to commands and contexts
